+++ src/Exchange.js
@@ -62,6 +62,12 @@ Object.defineProperty(ExchangeBase.prototype, 'stopped', {
   },
 });
 
+Object.defineProperty(ExchangeBase.prototype, 'undeliveredCount', {
+  get() {
+    return this[deliveryQueueSymbol].messageCount;
+  },
+});
+
 ExchangeBase.prototype.publish = function publish(routingKey, content, properties) {
   if (this[stoppedSymbol]) return;
   if (!this.bindingCount) return this[emitReturnSymbol](routingKey, content, properties);